Add Backup Library, Configuration, & Runner
Add a HsBackup module containing the types & functions for running the
backup service.

The Backup type defines a single remote location to be backed up, with
SSH credentials and various configuration data. The BackupRate type
represents the time delay between backups - Hourly, Daily, etc.

The AppState type keeps track of the current synchronization status of
the service - whether we have a queue of backups to make or if
everything is synced. The AppState has an `updateState` function that
serializes the state to a file whenever it is changed. We attempt to
read this file when starting the application so that interruptions like
reboots do not reset the backup queue.

The Config type wraps all these types up so their values are accessible
to the syncing & enqueueing threads.

The enqueueing thread simply checks the time every 5 minutes. If the
hour has changed, it enqueues the Hourly backups. If the day has
changed, it enqueues the Daily backups. If the Month has changed, it
enqueues the Monthly backups, etc.

The syncing thread reads the backup queue. If it is empty is simply
waits for 5 minutes and checks again. If there is a backup to process,
it runs the backup using rsync & restarts the process if an exception is
raised or rsync exits with an error. When a backup is completed, it is
removed from the queue and the sync thread continues to the next
enqueued backup.