Make Backups With `cp` When Possible
Modify the runBackup function to use `cp` to make a backup if we can.
This is possible when there is a higher-rate backup present. If so, we
simply copy that backup to the new backup directory. We use the `-l`
flag wil `cp` to make hardlinks instead of actually copying files, which
helps preserve diskspace.

Hourly backups will always use rsync, & daily backups will use rsync if
hourly backups are disabled.