Add Configuration Via YAML Files
Add a ConfigSpec type & readConfigFile function to the HsBackup module.
The ConfigSpec describes the format of the configuration files and is
created by passing a list of possible configuration files to the
readConfigFile function. The `run` function takes this ConfigSpec and
generates the Config type from it, which is then used as the environment
the syncing & enqueueing threads run in.

Modify the run function to check the file permissions of the log & state
paths passed via the ConfigSpec. If we cannot write to the log file, we
print a warning and log to stdout instead. If we cannot read and write
the state file, we create a file in the system's temporary directory and
print a warning that the program state will not be persisted across
reboots. We also ensure that all the backup folders are properly created
before kicking off the child threads.